Added Window console, Removed Main
Removed the main.py and added the winconsole.py to provide a console to the keylogger. The code of the main.py was shifted to key_convert.py and winconsole.py. At this stage the keylogger can log the alphabets, numbers, special characters, space key, enter key and backspace key. Any key other than the mentioned here will be ignored.